+++ src/types/index.ts
@@ -1,1 +1,2 @@
 export * from './rpc-response';
+export * from './rpc-body';